Move topic parsing logic into separate module to `parse_download_meta`
